+++ mrmustard/math/__init__.py
@@ -21,7 +21,6 @@ from .caching import *
 from .backend_base import *
 from .backend_manager import BackendManager
 from .backend_numpy import *
-from .compactFock import *
 from .lattice import *
 from .parameters import *
 from .parameter_set import *
